feat: Implementations of sends ICMP ECHO_REQUEST packets to network hosts
fping is a ping like program which uses the Internet Control Message Protocol (ICMP) echo request to determine if a target host is responding. fping differs from ping in that you can specify any number of targets on the command line, or specify a file containing the lists of targets to ping. Instead of sending to one target until it times out or replies, fping will send out a ping packet and move on to the next target in a round-robin fashion.

Log: Implementations of sends ICMP ECHO_REQUEST packets to network hosts

Issue: https://github.com/deepin-community/sig-deepin-sysdev-team/issues/182